Reading and writing fully supported
